mod promotion for Token struct, structure refactoring for auth module
